PYIC-5875: Move config to config dir
This puts all the config in once place, and makes it clear which config
is specifically for cucumber.

Also resets the defaults for the build environment to the correct values
