+++ api-tests/src/config/constants.ts
@@ -1,0 +1,13 @@
+export const BUILD_ENVIRONMENT = "build";
+export const BUILD_CORE_BACK_COMPONENT_ID =
+  "https://identity.build.account.gov.uk";
+export const BUILD_CORE_BACK_INTERNAL_API_URL =
+  "https://internal-test-api.identity.build.account.gov.uk";
+export const BUILD_CORE_BACK_EXTERNAL_API_URL =
+  "https://api.identity.build.account.gov.uk";
+export const BUILD_CORE_BACK_PUBLIC_ENCRYPTION_KEY =
+  '{"kty":"RSA","e":"AQAB","n":"nel7ibmSTaXWhwEAdqKTiEVcxsYgv6CdXaz90aVN7IorlaCeNj0j06OsA4zdmWEjj21wEZULsxPoZo5N_tsQ7NtOnOkcnDc-g_Nbpt0jelzJSbFRkx3kwXy8YIYKR_myNbiHNTTc7S6GkQRg0N1MPWtzoEKYJs41AN4onrsvUzgpCypWwPy2-ppsaDvms_11YA7A7x3zHj9oKCPJ_uk_0MV3vZAxCxbiPb9ABGWcoGQ5QKGfv40ylBsEdOhE3w-3SAAQIrrHyMRGGiPxcNO161XVL-lOnYt93FgEe16LgpfE22UdENfHnG0UQaTgph1Dm24oqn7qpPTY2DfER5HCKQ"}'; // pragma: allowlist secret
+export const ORCHESTRATOR_REDIRECT_URI =
+  "https://orch.stubs.account.gov.uk/callback";
+export const BUILD_JAR_SIGNING_KEY =
+  '{"kty":"EC","d":"OXt0P05ZsQcK7eYusgIPsqZdaBCIJiW4imwUtnaAthU","crv":"P-256","x":"E9ZzuOoqcVU4pVB9rpmTzezjyOPRlOmPGJHKi8RSlIM","y":"KlTMZthHZUkYz5AleTQ8jff0TJiS3q2OB9L5Fw4xA04"}'; // pragma: allowlist secret
