--- api-tests/src/constants.ts
@@ -1,13 +1,0 @@
-export const BUILD_ENVIRONMENT = "dev01";
-export const BUILD_CORE_BACK_COMPONENT_ID =
-  "https://dev-chrisw.01.dev.identity.account.gov.uk";
-export const BUILD_CORE_BACK_INTERNAL_API_URL =
-  "https://internal-api-dev-chrisw.01.dev.identity.account.gov.uk";
-export const BUILD_CORE_BACK_EXTERNAL_API_URL =
-  "https://api-dev-chrisw.01.dev.identity.account.gov.uk";
-export const BUILD_CORE_BACK_PUBLIC_ENCRYPTION_KEY =
-  '{"kty":"RSA","e":"AQAB","kid":"b454ac07-e188-415d-a3c8-f1d0d38aaecd","n":"loHeaSxvMgiHStKmb-ZK5ZPpwRWrhSSQ-nTyuKQj-mYWYFNGgGGNP-37Zvzo453bUGtEeFu1zdlLAoHyT3kgs1XdqXCvPinNccpJ8lWGXcFKGRhj5jxIiIMvEBHfLs_-cMIWW0166ndTT93ocoXdXaP64mH2iF7WWDyKqOcrVjuaUnbFbS4X2fhJwwRPj_Kin5jpJCx3MJd9eIuYyJB4CltbLTpX25oCwLw9t-p2lzHfazJSITcfTzEbOZV40fPJIR6HlJi7ApXYfAQ-dlbjMsYinFQnY6ILJXkbsjD4JXWUYaB0RbK8WTTKyehFU7P_Q8vFb7qWU4Xj9MTEHc7W3Q"}'; // pragma: allowlist secret
-export const ORCHESTRATOR_REDIRECT_URI =
-  "https://orch-dev-chrisw.01.core.dev.stubs.account.gov.uk/callback";
-export const BUILD_JAR_SIGNING_KEY =
-  '{"kty":"EC","d":"OXt0P05ZsQcK7eYusgIPsqZdaBCIJiW4imwUtnaAthU","crv":"P-256","x":"E9ZzuOoqcVU4pVB9rpmTzezjyOPRlOmPGJHKi8RSlIM","y":"KlTMZthHZUkYz5AleTQ8jff0TJiS3q2OB9L5Fw4xA04"}'; // pragma: allowlist secret
